requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Scheduling
apscheduler==3.10.4

# HTTP Client
httpx==0.26.0

# Email
aiosmtplib==3.0.1

# Templating
jinja2==3.1.3

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1

# Utilities
python-dateutil==2.8.2
croniter==2.0.1
